Marcia's mods, incorporating Jim's requests.
Marcia is having trouble with git, so I'm committing these changes.  I
also had to go in and change the title back from "...45m" to "...45M"
and "April 7" to "April 7th".

We need to extract the common stuff into a partial, but getting Marcia's
git working is also a high priority.