Got rid of debugger for now because debugger-ruby_core_source doesn't work.
The weird thing is that I did successfully install debugger using
ruby-1.9.3-p362, but although it worked on my laptop, it didn't work
on Brooks's, nor did it work on my imac.  I then uninstalled
ruby-1.9.3-p362 and reinstalled it and it failed to work again.

My suspcion is there's a difference in what command line was given to rvm
to install it, but it shouldn't matter too much because presumably
debugger-ruby_core_source will be updated soon and then we can just use that.