Replace update_protected_attrs_from_params with built in functionality
This method doesn't seem to have any reason to exist -- It provides
functionality that is built into attr_accessible (or with the
strong_params gem which is available now and replaces attr_accessible in
Rails 4).

The method was obtuse enough to require documentation nearly 3 times as
long as the actual method, and the code in the model and controller is
the same length without it.

Not sure what the reasoning was for including it, but it definitely has
no reason to exist